pyproject.toml
[project]
name = "mcp-python-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.1",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "ollama>=0.4.4",
    "httpx>=0.28.1",
    "rich>=13.10.5",
    "python-dotenv>=1.0.1",
    "websockets>=14.1",
    "aiohttp>=3.11.11",
    "aiofiles>=24.1.0",
]
